+++ src/Executors/Adapters/DbalAdapter.php
@@ -44,6 +44,11 @@ class DbalAdapter implements QueryExecutor
         }
     }
 
+    public function getConnection(): Connection
+    {
+        return $this->conn;
+    }
+
     public function compile(Query $query, ValueBinder $binder): string
     {
         return $this->compiler->compile($query, $binder);