+++ src/Executors/ExecutableUpdateQuery.php
@@ -5,6 +5,9 @@ namespace Somnambulist\Components\QueryBuilder\Executors;
 use Somnambulist\Components\QueryBuilder\Executors\Behaviours\ExecutableQuery;
 use Somnambulist\Components\QueryBuilder\Query\Type\UpdateQuery;
 
+/**
+ * @experimental Query execution is an experimental feature and may be removed.
+ */
 class ExecutableUpdateQuery extends UpdateQuery
 {
     use ExecutableQuery;